Fixed shebang patching in `mkNpmTar`
